% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_get.R
\name{dwd_get}
\alias{dwd_get}
\title{Download binary RADOLAN files from DWD Climate Data Centre server - BUG: CHECK HOW RECENT-HIST FLIP DATE (MONTH) IS DEFINED BY DWD}
\usage{
dwd_get(start, stop, dir, resolution = "hourly")
}
\arguments{
\item{start}{\code{POSIXct} value, start time from which on the data shall
be downloaded. If no POSIXct format is provided, the function will try to
convert the text string to POSIXct, assuming UTC as underlying time zone.}

\item{stop}{\code{POSIXct} value, stop time unti which on the data shall
be downloaded. If no POSIXct format is provided, the function will try to
convert the text string to POSIXct, assuming UTC as underlying time zone.}

\item{dir}{\code{Character} value, path to directory where the downloaded
binary files will be stored.}

\item{resolution}{\code{Character} values, keyword defining the temporal
resolution of the data to download. Possible choices are \code{"hourly"}
and \code{"5-min"}. Default value is \code{"hourly"}.}
}
\value{
Files written to a local directory.
}
\description{
The function manages the download of RADOLAN gridded precipitation raster
data sets from the DWD Climate Data Centre based on user defined start and
stop times and saves them as unpacked binary files in a user defined output
directory. Currently, only hourly and 5-min RADOLAN files are supported.
}
\examples{

\dontrun{

  dwd_get(start = as.POSIXct("2022-01-28"),
          stop = as.POSIXct("2022-02-02"),
          dir = "~/Downloads/RADOLAN/daily/")

}

}
\author{
Michael Dietze
}
