% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_subset.R
\name{dwd_subset}
\alias{dwd_subset}
\title{Extract subsets or time series from local binary RADOLAN files}
\usage{
dwd_subset(start, stop, dir, aoi, aggregate, cpu)
}
\arguments{
\item{start}{\code{POSIXct} value, start time from which on the data shall
be examined If no POSIXct format is provided, the function will try to
convert the text string to POSIXct, assuming UTC as underlying time zone.}

\item{stop}{\code{POSIXct} value, stop time unti which on the data shall
be examined If no POSIXct format is provided, the function will try to
convert the text string to POSIXct, assuming UTC as underlying time zone.}

\item{dir}{\code{Character} value, path to directory where the downloaded
binary files will be stored.}

\item{aoi}{Either a \code{simple feature} vector polygon or point object,
a \code{numeric} vector of length two with point coordinates, or a
\code{numeric} vector of length four with the boundary coordinates of a
rectangular area of interest (in the order \code{x_min}, \code{x_max},
\code{y_min}, \code{y_max}). All coordinates must be provided in the native
RADOLAN coordinate system.}

\item{aggregate}{\code{Character} value, optional keyword defining the
aggregation operation to perform with the extracted subset. If omitted,
the function will return a list of cropped terra rasters. Meaningful
keywords may be \code{"mean"}, \code{"median"}, \code{"sum"}, \code{"var"},
\code{"sd"}.}

\item{cpu}{\code{Numeric} value, fraction of CPUs to use for parallel
processing. If omitted, one CPU is used.}
}
\value{
A list with case specific elements. There will always be one list
element \code{time} that contains the time stamp of the respective subset.
If no aggregation function keyword (\code{fun}) is provided, the second
list element will contain the extracted values, either as a numeric matrix
(if aoi is a simple feature or boundary box), or as a numeric vector (if
aoi is a xy coordinate). If an aggregation keyword is provided, the second
list element will be a numeric vector containing the respective aggregated
values.
}
\description{
The function allows to extract either precipitation time series from
points or spatially extensive areas of interest, or cropping the RADOLAN
raster data to single pixels or areas of interest.
}
\examples{

\dontrun{

  rdl <- dwd_subset(start = as.POSIXct("2022-01-29 12:00:00"),
                    stop = as.POSIXct("2022-01-29 18:00:00"),
                    dir = "~/Downloads/RADOLAN/daily/",
                    aoi = c(576000, 578000, 5712000, 5715000))

}

}
\author{
Michael Dietze
}
