% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebergoetzen-package.R
\docType{package}
\name{eseis}
\alias{eseis}
\alias{ebergoetzen-package}
\title{ebergoetzen: Data Management for the Ebergoetzen Terra Network}
\description{
Functions to work with the Ebergoetzen Forest Site.
}
\details{
\tabular{ll}{
**Package:** \tab ebergoetzen \cr
**Type:** \tab Package \cr
**Version:** \tab 0.1.2 \cr
**Date:** \tab 2023-07-05 \cr
**License:** \tab GPL-3 \cr
}
}
\author{
Michael Dietze
}
\keyword{package}
