% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rainevents.R
\name{get_rainevents}
\alias{get_rainevents}
\title{Pick rain events from a time series}
\usage{
get_rainevents(
  data,
  intensity_min,
  duration_min = 0,
  amount_min = 0,
  pause_min = 0,
  lagtime = 0
)
}
\arguments{
\item{data}{Either a two-column \code{data.frame} with the time stamps
as \code{POSIXct} objects in the first and the \code{numeric} rain values
in the second column, or an EberGoetzenData (\code{egd}) object, i.e. a
nested list with data from the Ebergoetzen measurement site. See function
\code{ebergoetzen::read_data} for details.}

\item{intensity_min}{\code{Numeric} value, minimum rain intensity (amount
per time step) threshold to define a rain event.}

\item{duration_min}{\code{Numeric} value, minimum duration of consecutive
rainy conditions, in seconds. If omitted, the criterion is not applied.}

\item{amount_min}{\code{Numeric} value, minimum cumulative rain amount. If
omitted, the criterion is not applied.}

\item{pause_min}{\code{Numeric} value, minimum time span between two
consecutive rain events. The respective second event will be removed if
it occurs within the specified minimum pause. If omitted, the criterion
will not be applied.}

\item{lagtime}{\code{Numeric} value, lag time after the first occurrence
of zero rain after which subsequent rain is allowed to extend the duration
of the rain event. If omitted, the criterion will not be applied.}
}
\value{
A \code{data.frame} with identified rain events. Each event is
described by its start time, stop time, duration (in seconds), rain sum,
maximum and mean intensity.
}
\description{
The function screens a precipitation time series for rain events and
applies a series of rejection criteria to reduce the number of unwanted
events.
}
\examples{

## define example data path
dir_data <- paste0(system.file("extdata", package="ebergoetzen"), "/")

## read one day of EG data
x <- read_data(start = "2023-01-01",
               duration = 24 * 3600,
               dir = dir_data)

e <- get_rainevents(data = x,
                    lagtime = 1 * 3600,
                    pause_min = 2 * 3600,
                    intensity_min = 0.05,
                    duration_min = 1800,
                    amount_min = 0.1)

plot(x$time, x$meteo$precipitation$weather$d00$pluvio, type = "l")
abline(v = e$start, col = 3)
abline(v = e$stop, col = 2)
print(e)

}
\author{
Michael Dietze
}
