% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Load event data from Ebergoetzen data set}
\usage{
read_data(
  start,
  stop,
  duration,
  dir,
  average = FALSE,
  aggregate,
  smoothen,
  cum2rate = TRUE,
  egd = TRUE,
  silent = FALSE
)
}
\arguments{
\item{start}{\code{Character} value, start time of the data to import. If
lazy users only submit a text string instead of a POSIXct object, the
function will try to convert that text string.}

\item{stop}{\code{Character} value, stop time of the data to import. If
lazy users only submit a text string instead of a POSIXct object, the
function will try to convert that text string.}

\item{duration}{\code{Numeric} value, alternative to definition of
\code{stop} argument, duration of the event for which the data will be
imported, in seconds.}

\item{dir}{\code{Character} value, path to data directory.}

\item{average}{\code{Logical} value, option to average redundant
sensor data. Default is \code{FALSE}. Only works when the output is
an \code{egd} (EberGoetzenData) object, i.e. \code{egd = TRUE}.}

\item{aggregate}{\code{Numeric} value, factor by which to aggregate the
data set of initially 15 min sampling interval. If omitted, no aggregation
will be applied.}

\item{smoothen}{\code{Numeric} value, factor by which to smoothen the
data by running mean filter. If omitted, no smoothing is performed.
Option requires package caTools to be installed.}

\item{cum2rate}{\code{Logical} value, option to convert cumulative
precipitation values to rates (first derivative) Default is \code{TRUE}.}

\item{egd}{\code{Logical} value, option to return the data organised as
\code{egd} (EberGoetzenData) object. Default is \code{TRUE}. If set to
\code{FALSE}, a data frame will be returned, instead.}

\item{silent}{\code{Logical} value, option to suppress warnings, for
example of missing input data. Default is \code{FALSE}.}
}
\value{
Either an \code{egd} object or a \code{data.frame} with data
from the Ebergoetzen site for the defined event time interval.
}
\description{
The function reads the full data set of an event, defined by start time
and duration. It assumes the data is organised in directories by year and
as daily files.
}
\examples{

\dontrun{

## define example data path
dir_data <- paste0(system.file("extdata", package="ebergoetzen"), "/")

## read one day of EG data
x <- read_data(start = "2023-01-01",
               duration = 24 * 3600,
               dir = dir_data)

plot(x$time, x$meteo$temperature$north$hydroclip, type = "l")
}

}
\author{
Michael Dietze
}
