% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data.R
\name{select_data}
\alias{select_data}
\title{Select items of a data set}
\usage{
select_data(data, theme, metric, site, depth, simplify = TRUE)
}
\arguments{
\item{data}{\code{egd} object, see \code{ebergoetzen::read_data} for how to
create this object}

\item{theme}{\code{Character} vector, one or more keywords that describe
the data theme to be selected. Available themes are \code{"meteo"},
\code{"soil"} and \code{"tree"}. If omitted, all three themes will be
selected and returned as a list object.}

\item{metric}{\code{Character} vector, one or more keywords that describe
the metric (measured parameter) to be selected. Available themes are
\code{"temperature"}, \code{"precipitation"}, \code{"humidity"},
\code{"snow"}, \code{"radiation"}, \code{"wind"}, \code{"moisture"},
\code{"potential"}, \code{"throughfall"} and \code{"stemflow"}. Note that
some keywords apply for different \code{themes} and will be applied to
these themes if they are selected.}

\item{site}{\code{Character} vector, one or more keywords that describe
the site to be selected. Available sites are \code{"weather"},
\code{"north"} and \code{"south"}. If omitted, all three sites will be
selected and returned as a list object, if metrics exist in them.}

\item{depth}{\code{Character} vector, one or more keywords that describe
the depth to be selected. Available dpeths are \code{"d00"} (surface or
above ground sensor), \code{"d10"} (10 cm), \code{"d20"} (20 cm),
\code{"d30"} (30 cm), \code{"d50"} (50 cm) and \code{"d70"} (70 cm).
Note that for convenience, also just numeric values can be provided as
depth values (e.g., \code{c(0, 10, 20)}). If omitted, all depths will be
selected and returned as a list object, if metrics exist for them.}

\item{simplify}{\code{Logical} value, option to simplify output object
structure, i.e., all list elements that contain only a single element
will be removed. Default is \code{TRUE}.}
}
\description{
The function allows to isolate itmes of interest from a \code{egd} object.
}
\author{
Michael Dietze
}
